Methods

InverseLangevinApproximations.BergstromApproximation

InverseLangevinApproximations.CohenExact3_2

InverseLangevinApproximations.CohenRounded3_2

InverseLangevinApproximations.DarabiItskov

InverseLangevinApproximations.Jedynak2017

InverseLangevinApproximations.KuhnGrunApproximation

InverseLangevinApproximations.NguessongBedaPeyraut

InverseLangevinApproximations.PadeApproximation3_2

InverseLangevinApproximations.PadeApproximation_1_2

InverseLangevinApproximations.PadeApproximation_1_4

InverseLangevinApproximations.PadeApproximation_3_0

InverseLangevinApproximations.PadeApproximation_5_0

InverseLangevinApproximations.PusoApproximation

InverseLangevinApproximations.TreloarApproximation

InverseLangevinApproximations.WarnerApproximation

Description

InverseLangevinApproximations.BergstromApproximation — Type

From Ref. [1]:

\[\mathcal{L}^{-1}(y) \approx\begin{cases} 1.31446\tan{1.58986y}+0.91209y & |y|\leq 0.84136 \\ \frac{1}{\text{sign}(y)-y} & 0.84136 \leq |x| < 1.0 \\ \end{cases}\]

InverseLangevinApproximations.CohenExact3_2 — Type

From Ref. [2]:

\[\mathcal{L}^{-1}(y) \approx y\frac{3-\frac{36}{35}y^2}{1-\frac{33}{35}y^2}\]

InverseLangevinApproximations.CohenRounded3_2 — Type

From Ref. [2]:

\[\mathcal{L}^{-1}(y) \approx y\frac{3-y^2}{1-y^2}\]

InverseLangevinApproximations.DarabiItskov — Type

From Ref. [3]:

\[{\mathcal{L}^{- 1}}(y) \approx y\frac{{y^{2} - 3y + 3}}{{1 - y}}.\]

InverseLangevinApproximations.Jedynak2017 — Type

From Ref. [4]:

\[\mathcal{L}^{-1}(y) \approx y\frac{3-\frac{773}{768}y^2-\frac{1300}{1351}y^4+\frac{501}{340}y^6-\frac{678}{138}y^8}{(1-y)(1+\frac{866}{853}*y)}\]

InverseLangevinApproximations.KuhnGrunApproximation — Type

From Ref. [5]:

\[\mathcal{L}^{-1}(y) \approx 3y + \frac{9}{5}y^3 + \frac{297}{175}y^5 + \frac{1539}{875}y^7 + \frac{126117}{67375}y^9 + \frac{43733439}{21896875}y^{11} + \frac{231321177}{109484375}y^{13} + \frac{20495009043}{9306171875}y^{15} + \frac{1073585186448381}{476522530859375}y^{17} + \frac{4387445039583}{1944989921875}y^{19}\]

InverseLangevinApproximations.NguessongBedaPeyraut — Type

From Ref. [6]:

\[{\mathcal{L}^{-1}}\left( y \right)\approx y\frac{3-y^{2}}{1-y^{2}}-0.488y^{3.243}+3.311y^{4.789}\left( y-0.76 \right)\left( y-1 \right)\]

InverseLangevinApproximations.PadeApproximation3_2 — Type

From Ref. [2]:

\[\mathcal{L}^{-1}(y) \approx y \frac{(3 - \frac{36}{35} y^2)}{(1 - \frac{33}{35} y^2)}\]

InverseLangevinApproximations.PadeApproximation_1_2 — Type

From Ref. [7]:

\[\mathcal{L}^{-1}(y) \approx \frac{3y}{1-\frac{3}{5}y^2}\]

InverseLangevinApproximations.PadeApproximation_1_4 — Type

From Ref. [7]:

\[\mathcal{L}^{-1}(y) \approx \frac{3y}{1-\frac{3}{5}y^2-\frac{36}{175}y^4}\]

InverseLangevinApproximations.PadeApproximation_3_0 — Type

From Ref. [7]:

\[\mathcal{L}^{-1}(y) \approx 3y + \frac{9}{5}y^3\]

InverseLangevinApproximations.PadeApproximation_5_0 — Type

From Ref. [7]:

\[\mathcal{L}^{-1}(y) \approx 3y + \frac{9}{5}y^3 + \frac{297}{175}y^5\]

InverseLangevinApproximations.PusoApproximation — Type

From Ref. [8]:

\[\mathcal{L}^{-1}(y) \approx \frac{3y}{1-y^3}\]

InverseLangevinApproximations.TreloarApproximation — Type

From Ref. [9]:

\[\mathcal{L}^{-1}(y) \approx \frac{3 y}{(1 - (\frac{3}{5} y^2 + \frac{36}{175} y^4 + \frac{108}{875} y^6))}\]

InverseLangevinApproximations.WarnerApproximation — Type

From Ref. [10]:

\[\mathcal{L}^{-1}(y) \approx \frac{3y}{1-y^2}\]

References

[1]

J. S. Bergstr{\"o}m. Large strain time-dependent behavior of elastomeric materials. Phd thesis, Massachusetts Institute of Technology (1999).

[2]

A. Cohen. A Pad{\'e} approximant to the inverse Langevin function. Rheologica acta 30, 270–273 (1991).

[3]

E. Darabi and M. Itskov. A simple and accurate approximation of the inverse Langevin function. Rheologica Acta 54, 455–459 (2015).

[4]

R. Jedynak. New facts concerning the approximation of the inverse Langevin function. Journal of Non-Newtonian Fluid Mechanics 249, 8–25 (2017).

[5]

W. Kuhn and F. Gr{\"u}n. Beziehungen zwischen elastischen Konstanten und Dehnungsdoppelbrechung hochelastischer Stoffe. Kolloid-Zeitschrift 101, 248–271 (1942).

[6]

A. N. Nguessong, T. Beda and F. Peyraut. A new based error approach to approximate the inverse Langevin function. Rheologica Acta 53, 585–591 (2014).

[7]

R. Jedynak. Approximation of the inverse Langevin function revisited. Rheologica Acta 54, 29–39 (2015).

[8]

M. Puso. Mechanistic constitutive models for rubber elasticity and viscoelasticity. Lawrence Livermore National Lab.(LLNL), Livermore, CA (United States) (2003).

[9]

L. R. Treloar. The mechanics of rubber elasticity. Proceedings of the Royal Society of London. A. Mathematical and Physical Sciences 351, 301–330 (1976).

[10]

H. R. Warner Jr. Kinetic theory and rheology of dilute suspensions of finitely extendible dumbbells. Industrial \& Engineering Chemistry Fundamentals 11, 379–387 (1972).